test(live): the guard accepted a call that never runs, and the test was misfiled (#278)
Round 3. Smaller findings than the first two rounds, but the guard's docstring was
claiming more than it enforced.

**`ast.walk` accepts unreachable calls.** The guard said the body "really CALLS"
`self.observer.reset()`, and it correctly rejected a comment, a docstring and a string
literal -- but accepted `if False: self.observer.reset()`, a nested `def` that never
runs, and a statement after `return`. It scans top-level statements now, in the three
shapes a call takes there (bare, returned, assigned -- the SLTP envs assign their
`super()._reset(...)`). Mutation-checked against both unreachable forms.

**`test_a_live_account_keeps_its_run_level_bankruptcy_baseline` passes on main** and on
this branch with the fix mutated out, so it is a guard against re-introducing round 1's
defect rather than a test of this PR. Worth keeping -- round 1 made exactly that mistake
and it deletes the cross-episode drawdown circuit breaker -- but it exercises
AlpacaTorchTradingEnv with mock observers and has nothing to do with replay. Moved to
test_live_env_base.py, where the other cross-venue live contracts live.

Prose trimmed to match the fix: production is now 8 executable lines. The 3-line
ordering comment is stated once and pointed at from the other four sites, and the two
standalone observation classes point at the base's docstring rather than restating it.
242 lines from 248, of which the repeated block is gone.

Two corrections carried from the review: this file has FIVE older AST guards, not four
(the round 2 message said four), and fb10dfc's "3648 passed" was a4c201f's number
carried forward -- it was 3650. Third factual error in this PR's messages, so I am
listing them rather than quietly fixing them.

Noted for whoever reads a red suite: `tests/actor/test_frontier_llm_actor.py::
test_generate_batch_preserves_order` is a pre-existing flake, failing 2 of 3 full runs
on main and passing in isolation. Not this branch.

Full suite 3650 passed, 0 xfailed.
