+++ tests/envs/replay/test_reset_rewinds_278.py
@@ -76,46 +76,3 @@ def test_a_second_episode_rewinds_the_data_the_balance_and_the_baseline():
     # No baseline assertion here: replay rewinds the balance to the SAME number, so
     # capturing once and capturing per episode are indistinguishable in this scenario.
     # An assertion that cannot fail is worse than none -- the sibling test carries it.
-
-
-def test_a_live_account_keeps_its_run_level_bankruptcy_baseline():
-    """The yardstick must not chase the account down.
-
-    Rebasing the baseline every episode looks like offline parity and is not: offline
-    resets the BALANCE and the baseline together, so its ratio starts at 1.0 because the
-    account was reset too. A live account persists, so rebasing only the yardstick
-    removes the cross-episode drawdown circuit breaker entirely -- an account halving
-    every episode reached 0.39% of its starting value without ever reporting bankrupt.
-    """
-    from tests.mocks.alpaca import MockObserver, MockTrader
-    from torchtrade.envs.live.alpaca.env import (
-        AlpacaTorchTradingEnv,
-        AlpacaTradingEnvConfig,
-    )
-
-    trader = MockTrader(initial_cash=1000.0)
-    env = AlpacaTorchTradingEnv(
-        config=AlpacaTradingEnvConfig(
-            symbol="BTC/USD", window_sizes=[10], bankrupt_threshold=0.1,
-        ),
-        observer=MockObserver(window_sizes=[10]),
-        trader=trader,
-    )
-    env._wait_for_next_timestamp = lambda: None
-
-    env.reset()
-    assert env.initial_portfolio_value == pytest.approx(1000.0)
-
-    # The account halves each episode. A live observer rewinds nothing, so the baseline
-    # must stay at the equity the RUN started from.
-    for equity in (500.0, 250.0, 125.0, 62.5):
-        trader.cash = equity
-        env.reset()
-        assert env.initial_portfolio_value == pytest.approx(1000.0), (
-            f"baseline rebased to {env.initial_portfolio_value} at equity {equity}; it "
-            f"now chases the account down and bankruptcy can never fire (#278)"
-        )
-
-    assert env._check_termination(62.5) is True, (
-        "62.5 is below 10% of the 1000 this run started with and must terminate"
-    )
