+++ torchtrade/envs/live/binance/observation.py
@@ -13,13 +13,8 @@ class BinanceObservationClass:
     Supports multiple timeframes and custom feature preprocessing.
     """
 
-
     def reset(self) -> None:
-        """Rewind to the start of an episode; a live observer has nothing to rewind.
-
-        See BaseFuturesObservationClass.reset for why this exists and why it does not
-        touch the bankruptcy baseline (#278).
-        """
+        """No-op; see BaseFuturesObservationClass.reset (#278)."""
 
     def __init__(
         self,
